Ryddet i validation og index fil
