+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/partials/adresser/adressetyper/index.ts
@@ -1,0 +1,21 @@
+import { UtenlandskAdresse } from './UtenlandskAdresse'
+import { UkjentBosted } from './UkjentBosted'
+import { Vegadresse } from './Vegadresse'
+import { VegadresseVelger } from './VegadresseVelger'
+import { Matrikkeladresse } from './Matrikkeladresse'
+import { MatrikkeladresseVelger } from './MatrikkeladresseVelger'
+import { MatrikkeladresseTilfeldig } from './MatrikkeladresseTilfeldig'
+import { OppholdAnnetSted } from './OppholdAnnetSted'
+import { Postboksadresse } from './Postboksadresse'
+
+export {
+	UtenlandskAdresse,
+	UkjentBosted,
+	Vegadresse,
+	VegadresseVelger,
+	Matrikkeladresse,
+	MatrikkeladresseVelger,
+	MatrikkeladresseTilfeldig,
+	OppholdAnnetSted,
+	Postboksadresse,
+}
