+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/partials/identifikasjon/validation.js
@@ -1,0 +1,29 @@
+import * as Yup from 'yup'
+import { requiredString } from '~/utils/YupValidations'
+
+const personnavnSchema = Yup.object({
+	fornavn: Yup.string(),
+	mellomnavn: Yup.string(),
+	etternavn: Yup.string(),
+})
+
+export const falskIdentitet = Yup.array().of(
+	Yup.object({
+		rettIdentErUkjent: Yup.boolean(),
+		rettIdentitetVedIdentifikasjonsnummer: Yup.string().nullable(),
+		rettIdentitetVedOpplysninger: Yup.object({
+			foedselsdato: Yup.string().nullable(),
+			kjoenn: Yup.string().nullable(),
+			personnavn: personnavnSchema.nullable(),
+			statsborgerskap: Yup.array().of(Yup.string()),
+		}),
+	})
+)
+
+export const utenlandskId = Yup.array().of(
+	Yup.object({
+		identifikasjonsnummer: requiredString,
+		opphoert: requiredString,
+		utstederland: requiredString,
+	})
+)
